frontend/requirements.txt
Flask==2.3.2
python-dotenv
pydub
sqlalchemy
databricks-sql-connector
git+https://github.com/sebrahimi1988/databricks-model-serving
